# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "EBOS API",
        "contact": {
            "name": "Support Team",
            "email": "support@example.com"
        },
        "version": "0.0.0"
    },
    "paths": {
        "/api/v1/barangays": {
            "get": {
                "tags": [
                    "Barangay"
                ],
                "summary": "Get all barangays",
                "operationId": "getBarangays",
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "current_page": {
                                            "format": "string",
                                            "example": 1
                                        },
                                        "data": {
                                            "type": "array",
                                            "format": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "format": "uuid",
                                                        "example": "1"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "first_page_url": {
                                            "format": "string",
                                            "example": "http://localhost:8000/api/v1/barangays?page=1"
                                        },
                                        "from": {
                                            "format": "string",
                                            "example": null
                                        },
                                        "last_page": {
                                            "format": "integer",
                                            "example": 1
                                        },
                                        "last_page_url": {
                                            "format": "string",
                                            "example": "http://localhost:8000/api/v1/barangays?page=1"
                                        },
                                        "next_page_url": {
                                            "format": "string",
                                            "example": null
                                        },
                                        "per_page": {
                                            "format": "integer",
                                            "example": 20
                                        },
                                        "total": {
                                            "format": "integer",
                                            "example": 0
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "$ref": "#/components/responses/notauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Barangay"
                ],
                "summary": "Create a Barangay",
                "operationId": "dd03c7b7f1cff987c39f2cf10af83256",
                "requestBody": {
                    "description": "Pass barangay information",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Angels City"
                                    },
                                    "address": {
                                        "type": "string",
                                        "example": "123 Street"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successfully Added Barangay.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "format": "string",
                                            "example": "success"
                                        },
                                        "message": {
                                            "format": "string",
                                            "example": "Complaints retrieved successfully."
                                        },
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "format": "uuid",
                                                    "example": "d0b0c8c0-5f5a-4a5a-9b1a-3c1b0f7f5f1a"
                                                },
                                                "name": {
                                                    "format": "string",
                                                    "example": "Angeles City"
                                                },
                                                "address": {
                                                    "format": "string",
                                                    "example": "123 Street"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/barangays/{barangay_id}": {
            "get": {
                "tags": [
                    "Barangay"
                ],
                "summary": "Get one barangay",
                "operationId": "5afa87e1fcd28eb0e602561afcdf9b6d",
                "parameters": [
                    {
                        "name": "barangay_id",
                        "in": "path",
                        "description": "Barangay ID",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "format": "string",
                                            "example": "success"
                                        },
                                        "message": {
                                            "format": "string",
                                            "example": "Barangay retrieved successfully."
                                        },
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "format": "uuid",
                                                    "example": "d0b0c8c0-5f5a-4a5a-9b1a-3c1b0f7f5f1a"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Barangay"
                ],
                "summary": "Update a Barangay",
                "operationId": "a7228dffe82a86c61393d91893fe89b3",
                "parameters": [
                    {
                        "name": "barangay_id",
                        "in": "path",
                        "description": "Barangay ID",
                        "required": true
                    }
                ],
                "requestBody": {
                    "description": "Pass barangay information",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Angels City"
                                    },
                                    "address": {
                                        "type": "string",
                                        "example": "123 Street"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successfully Updated Barangay.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "format": "string",
                                            "example": "success"
                                        },
                                        "message": {
                                            "format": "string",
                                            "example": "Complaints retrieved successfully."
                                        },
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "format": "uuid",
                                                    "example": "d0b0c8c0-5f5a-4a5a-9b1a-3c1b0f7f5f1a"
                                                },
                                                "name": {
                                                    "format": "string",
                                                    "example": "Angeles City"
                                                },
                                                "address": {
                                                    "format": "string",
                                                    "example": "123 Street"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "responses": {
            "notauthenticated": {
                "description": "Respone for Unauthenticated User",
                "content": {
                    "application/json": {
                        "schema": {
                            "properties": {
                                "code": {
                                    "format": "string",
                                    "example": "401"
                                },
                                "message": {
                                    "format": "string",
                                    "example": "You're unauthorized to make a request!"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            }
        },
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "bearerAuth",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}